Refactors modules code and organization
