+++ lib/src/modules/home/home_main/home_page.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
-import 'package:raro_budget/src/modules/home/pages/home_controller.dart';
-import 'package:raro_budget/src/modules/home/widgets/custom_card_dia_a_dia/custom_card_dia_a_dia.dart';
-import 'package:raro_budget/src/modules/home/widgets/custom_drawer/custom_drawer_widget.dart';
-import 'package:raro_budget/src/modules/home/widgets/custom_general_balance/custom_general_balance_widget.dart';
+import 'package:raro_budget/src/modules/home/home_main/custom_card_dia_a_dia/custom_card_dia_a_dia.dart';
+import 'package:raro_budget/src/modules/home/home_main/custom_drawer/custom_drawer_widget.dart';
+import 'package:raro_budget/src/modules/home/home_main/custom_general_balance/custom_general_balance_widget.dart';
+import 'package:raro_budget/src/modules/home/home_main/home_controller.dart';
+
 import 'package:raro_budget/src/modules/home/widgets/custom_last_transactions/custom_last_transactions_widget.dart';
 import 'package:raro_budget/src/shared/constants/app_text_styles.dart';
 import 'package:raro_budget/src/shared/widgets/custom_appbar/custom_appbar.dart';
