+++ lib/src/modules/home/home_module.dart
@@ -1,9 +1,11 @@
 import 'package:flutter_modular/flutter_modular.dart';
+import 'package:raro_budget/src/modules/home/home_filled/home_page_filled.dart';
+import 'package:raro_budget/src/modules/home/home_filled/home_page_filled_controller.dart';
 import 'package:raro_budget/src/modules/home/home_in_out/home_in_page.dart';
 import 'package:raro_budget/src/modules/home/home_in_out/home_out_page.dart';
-import 'package:raro_budget/src/modules/home/pages/home_page.dart';
-import 'package:raro_budget/src/modules/home/pages/home_page_filled.dart';
-import 'package:raro_budget/src/modules/home/pages/home_page_filled_controller.dart';
+import 'package:raro_budget/src/modules/home/home_main/home_controller.dart';
+import 'package:raro_budget/src/modules/home/home_main/home_page.dart';
+
 import 'package:raro_budget/src/shared/auth/auth_repository.dart';
 import 'package:raro_budget/src/shared/models/firebase_model.dart';
 
@@ -11,7 +13,7 @@ class HomeModule extends Module {
   @override
   List<Bind> get binds => [
         Bind.singleton((i) => HomePageFilledController(i.get<FirebaseModel>())),
-        //  Bind.singleton((i) => HomeController(i.get<FirebaseModel>())),
+        Bind.singleton((i) => HomeController(i.get<FirebaseModel>())),
         Bind.singleton((i) => FirebaseModel(i.get<AuthRepository>())),
       ];
